Colocate tool approval hooks in SWR
